Handle jwt expiries for API requests given in Clerk, tokens last only for 1 minute
https://twitter.com/tweetsbycolin/status/1644137146503286786

"Tons of auth services set the JWT expiration equal to the session lifetime, which can be days, weeks, or months. That makes urgent revocation impossible!"

"Clerk decouples session lifetime from JWT expiration.  Our JWTs only last a minute, and we use a poller on the frontend to refresh.
The refresh leverages both a database check (to check for session expiration or revocation) and an HttpOnly cookie (so it's resilient to XSS attacks)."